project/requirements.txt
Flask
Flask-SocketIO
paho-mqtt
requests
sqlite3
